add update and delete method in controller

+++ src/main/java/com/wcs/restcontrollercourse/controller/ArticleController.java
@@ -8,6 +8,7 @@ import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -83,9 +84,16 @@ public class ArticleController {
 	// Delete article
 	//// http://localhost:8080/articles/{id}
 	@DeleteMapping("/{id}")
-	public boolean delete(@PathVariable Long id) {
+	public ResponseEntity<?> deleteArticle(@PathVariable(required = true) Long id) {
+		Boolean exist = articleRepository.existsById(id);
+		// Si mon article n'xiste pas , je renvoie une erreur
+		if (!exist) {
+			throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+		}
+		// Si il existe je le supprime de la Base de Données
 		articleRepository.deleteById(id);
-		return true;
+		return new ResponseEntity<>(HttpStatus.OK);
+
 	}
 
 }
